Initial example of Adapter design pattern
Created the initial example for the Design Pattern called Adapter.
This Design Pattern belongs to the Structural Pattterns.

Still missing the unit tests... I don't put it together once I'm
looking for a new libraries to perform the assertions... I'm pretty
familiar with Junit and Mockito but I would love try a couple new
technologies for that.

I hope soon I send a new commit for that