Created the first unit tests
Added the first unit tests using a pretty nice library called assertJ
it seems that is very nice. At least the assertion is so powerful.

I will try to use it and may be in the next commits I will add the
travis CI job in order to keep the code always with no errors. This
is the idea ;)