Added a new design pattern. Facade
Created the project called facade. This new project will cover the
design pattern Facade. I hope it is going to be enough to explain
this Design Pattern, once it is so simple.

The idea behind it is get easy something complex for someone.
For example: if the user needs to start the server we don`t need
to force him/her to meet all the instructions or all the steps to
do that. So to get this task easy for him/her I`ve created a Facade
class to turn it simplest possible for user.