Adding a new Design Pattern. Composite
At this time I'd like to introduce the design pattern
called Composite.

This is a very important Desing Pattern, actually this one
is a Structural one. This is used whenever the developer
needs to solve a problem related to structure, like an HTML
page or XML file or like your company structure, as this
example shown.

I hope you have fun ;)