Created the design pattern Bridge
It is time, finally for another Design pattern
this time the pick one is Bridge.

Using it is possible to separate the concrete
implementation from the Abstract class, using
a reference to the implementer object.

It is very fun and very useful.

Also I've changed the travis CI script in order to
add the bridge project as well.