Added the design pattern Observer
For me the Observer Design patter is very nice
and it is very useful.

I put it here using an example of Real time news
where the users would like to keep sync with
some soccer match for example.

Also, I`ve changed the travis CI script in order
to take care of the observer project as well